Fuzzy Patcher/Matcher now avoids modifying previously applied patches.

+++ Patcher.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 
 namespace DiffPatch
@@ -52,6 +53,11 @@ namespace DiffPatch
 			
 			public WorkingPatch(Patch patch) : base(patch) {}
 
+			public LineRange? KeepoutRange1 => result?.appliedPatch?.TrimmedRange1;
+			public LineRange? KeepoutRange2 => result?.appliedPatch?.TrimmedRange2;
+
+			public int? AppliedDelta => result?.appliedPatch?.length2 - result?.appliedPatch?.length1;
+
 			public void Fail() {
 				result = new Result {patch = this, success = false};
 			}
@@ -93,14 +99,17 @@ namespace DiffPatch
 		private List<string> lines;
 		private bool applied;
 
-		//we maintain delta as the offset of the last patch (applied location - expected location)
-		//this way if a line is inserted, and all patches are offset by 1, only the first patch is reported as offset
+		// Last here means highest line number, not necessarily most recent.
+		// Patches can only apply before lastAppliedPatch in fuzzy mode
+		private Patch lastAppliedPatch = null;
+
+		// we maintain delta as the offset of the last patch (applied location - expected location)
+		// this way if a line is inserted, and all patches are offset by 1, only the first patch is reported as offset
+		// normally this is equivalent to `lastAppliedPatch?.AppliedOffset` but if a patch fails, we subtract its length delta from the search offset
 		private int searchOffset;
-		//running tally of length2 - length1
-		//a target line in the currently patched file is patchedDelta ahead of the original file
-		private int patchedDelta;
-		//to prevent offset or fuzzy searching too far back
-		private int lastPatchedLine;
+
+		// patches applying within this range (due to fuzzy matching) will cause patch reordering
+		private LineRange ModifiedRange => new LineRange { start = 0, end = lastAppliedPatch?.TrimmedRange2.end ?? 0 };
 
 		private readonly CharRepresenter charRep;
 		private string lmText;
@@ -155,6 +164,9 @@ namespace DiffPatch
 		private Patch ApplyExactAt(int loc, WorkingPatch patch) {
 			if (!patch.ContextLines.SequenceEqual(lines.GetRange(loc, patch.length1)))
 				throw new Exception("Patch engine failure");
+			if (!CanApplySafelyAt(loc, patch))
+				throw new Exception("Patch affects another patch");
+
 
 			lines.RemoveRange(loc, patch.length1);
 			lines.InsertRange(loc, patch.PatchedLines);
@@ -169,6 +181,7 @@ namespace DiffPatch
 				wmLines.InsertRange(loc, patch.wmPatched);
 			}
 
+			int patchedDelta = patches.Where(p => p.KeepoutRange2?.end <= loc).Sum(p => p.AppliedDelta.Value);
 			Patch appliedPatch = patch;
 			if (appliedPatch.start2 != loc || appliedPatch.start1 != loc - patchedDelta)
 				appliedPatch = new Patch(patch) { //create a new patch with different applied position if necessary
@@ -176,12 +189,27 @@ namespace DiffPatch
 					start2 = loc
 				};
 
-			//update the patchedDelta and searchOffset
-			searchOffset = loc - patch.start2;
-			patchedDelta += patch.length2 - patch.length1;
-			lastPatchedLine = loc + patch.length2;
+			
+			// update the applied location for patches following this one in the file, but preceding it in the patch list
+			// can only happen if fuzzy matching causes a patch to move before one of the previously applied patches
+			if (loc < ModifiedRange.end) {
+				foreach (var p in patches.Where(p => p.KeepoutRange2?.start > loc))
+					p.result.appliedPatch.start2 += appliedPatch.length2 - appliedPatch.length1;
+			}
+			else {
+				lastAppliedPatch = appliedPatch;
+			}
 
+			searchOffset = appliedPatch.start2 - patch.start2;
 			return appliedPatch;
+		}
+
+		private bool CanApplySafelyAt(int loc, Patch patch) {
+			if (loc >= ModifiedRange.end)
+				return true;
+
+			var range = new LineRange { start = loc, length = patch.length1 };
+			return patches.All(p => !p.KeepoutRange2?.Contains(range) ?? true);
 		}
 
 		private bool ApplyExact(WorkingPatch patch) {
@@ -209,7 +237,10 @@ namespace DiffPatch
 
 			int forward = lmText.IndexOf(patch.lmContext, loc, StringComparison.Ordinal);
 			int reverse = lmText.LastIndexOf(patch.lmContext, Math.Min(loc+patch.lmContext.Length, lines.Count-1), StringComparison.Ordinal);
-			if (reverse < lastPatchedLine)
+
+			if (!CanApplySafelyAt(forward, patch))
+				forward = -1;
+			if (!CanApplySafelyAt(reverse, patch))
 				reverse = -1;
 
 			if (forward < 0 && reverse < 0)
@@ -230,12 +261,24 @@ namespace DiffPatch
 			if (loc + patch.length1 > wmLines.Count)//initialise search at end of file if loc is past file length
 				loc = wmLines.Count - patch.length1;
 
-			int[] match = FindMatch(loc, patch.wmContext, out float matchQuality);
+			(int[] match, float matchQuality) = FindMatch(loc, patch.wmContext);
 			if (match == null)
 				return false;
 
+			var fuzzyPatch = new WorkingPatch(AdjustPatchToMatchedLines(patch, match, lines));
+			if (wmLines != null) fuzzyPatch.WordsToChars(charRep);
+			if (lmText != null) fuzzyPatch.LinesToChars(charRep);
+
+			int at = match.First(i => i >= 0); //if the patch needs lines trimmed off it, the early match entries will be negative
+			patch.Succeed(Mode.FUZZY, ApplyExactAt(at, fuzzyPatch));
+			patch.AddOffsetResult(fuzzyPatch.start2 - loc, lines.Count);
+			patch.AddFuzzyResult(matchQuality);
+			return true;
+		}
+
+		public static Patch AdjustPatchToMatchedLines(Patch patch, int[] match, IReadOnlyList<string> lines) {
 			//replace the patch with a copy
-			var fuzzyPatch = new WorkingPatch(patch);
+			var fuzzyPatch = new Patch(patch);
 			var diffs = fuzzyPatch.diffs; //for convenience
 
 			//keep operations, but replace lines with lines in source text
@@ -243,7 +286,7 @@ namespace DiffPatch
 			//unmatched target lines (increasing offset) are added to the patch
 			for (int i = 0, j = 0, ploc = -1; i < patch.length1; i++) {
 				int mloc = match[i];
-				
+
 				//insert extra target lines into patch
 				if (mloc >= 0 && ploc >= 0 && mloc - ploc > 1) {
 					//delete an unmatched target line if the surrounding diffs are also DELETE, otherwise use it as context
@@ -267,54 +310,94 @@ namespace DiffPatch
 
 			//finish our new patch
 			fuzzyPatch.RecalculateLength();
-			if (wmLines != null) fuzzyPatch.WordsToChars(charRep);
-			if (lmText != null) fuzzyPatch.LinesToChars(charRep);
-
-			int at = match.First(i => i >= 0); //if the patch needs lines trimmed off it, the early match entries will be negative
-			patch.Succeed(Mode.FUZZY, ApplyExactAt(at, fuzzyPatch));
-			patch.AddOffsetResult(fuzzyPatch.start2 - loc, lines.Count);
-			patch.AddFuzzyResult(matchQuality);
-			return true;
+			return fuzzyPatch;
 		}
 
-		private int[] FindMatch(int loc, IReadOnlyList<string> wmContext, out float bestScore) {
-			bestScore = MinMatchScore;
+		private (int[] match, float score) FindMatch(int loc, IReadOnlyList<string> wmContext) {
+			// fuzzy matching is more complex because we need to split up the patched file to only search _between_ previously applied patches
+			var keepoutRanges = patches.Select(p => p.KeepoutRange2).Where(r => r != null).Select(r => r.Value);
+
+			// parts of file to search in
+			var ranges = new LineRange { length = wmLines.Count }.Except(keepoutRanges).ToArray();
+
+			return FuzzyMatch(wmContext, wmLines, loc, MaxMatchOffset, MinMatchScore, ranges);
+		}
+		
+		public static (int[] match, float score) FuzzyMatch(IReadOnlyList<string> wmPattern, IReadOnlyList<string> wmText, int loc, int maxMatchOffset = MatchMatrix.DefaultMaxOffset, float minMatchScore = FuzzyLineMatcher.DefaultMinMatchScore, LineRange[] ranges = default) {
+			if (ranges == null)
+				ranges = new LineRange[] { new LineRange { length = wmText.Count } };
+
+			// we're creating twice as many MatchMatrix objects as we need, incurring some wasted allocation and setup time, but it reads easier than trying to precompute all the edge cases
+			var fwdMatchers = ranges.Select(r => new MatchMatrix(wmPattern, wmText, maxMatchOffset, r)).SkipWhile(m => loc > m.WorkingRange.last).ToArray();
+			var revMatchers = ranges.Reverse().Select(r => new MatchMatrix(wmPattern, wmText, maxMatchOffset, r)).SkipWhile(m => loc < m.WorkingRange.first).ToArray();
+
+			int warnDist = OffsetWarnDistance(wmPattern.Count, wmText.Count);
+			float penaltyPerLine = 1f / (10*warnDist);
+
+			var fwd = new MatchRunner(loc, 1, fwdMatchers, penaltyPerLine);
+			var rev = new MatchRunner(loc,-1, revMatchers, penaltyPerLine);
+
+			float bestScore = minMatchScore;
 			int[] bestMatch = null;
+			while (fwd.Step(ref bestScore, ref bestMatch) | rev.Step(ref bestScore, ref bestMatch));
 
-			var mmForward = new MatchMatrix(wmContext, wmLines, MaxMatchOffset);
-			float score = mmForward.Init(loc);
-			if (score >= bestScore) {
-				bestScore = score;
-				bestMatch = mmForward.Path();
+			return (bestMatch, bestScore);
+		}
+
+		private struct MatchRunner
+		{
+			private int loc;
+			private readonly int dir;
+			private readonly MatchMatrix[] mms;
+			private readonly float penaltyPerLine;
+
+			// used as a Range/Slice for the MatchMatrix array
+			private LineRange active;
+			private float penalty;
+
+			public MatchRunner(int loc, int dir, MatchMatrix[] mms, float penaltyPerLine) {
+				this.loc = loc;
+				this.dir = dir;
+				this.mms = mms;
+				this.penaltyPerLine = penaltyPerLine;
+				active = new LineRange();
+				penalty = -0.1f; // start penalty at -10%, to give some room for finding the best match if it's not "too far"
 			}
 
-			var mmReverse = new MatchMatrix(wmContext, wmLines, MaxMatchOffset);
-			mmReverse.Init(loc);
+			public bool Step(ref float bestScore, ref int[] bestMatch) {
+				if (active.first == mms.Length)
+					return false;
 
-			int warnDist = OffsetWarnDistance(wmContext.Count, lines.Count);
-			for (int i = 0; mmForward.CanStepForward || mmReverse.CanStepBackward; i++) {
-				//within the warning range it's a straight up fight
-				//past the warning range, quality is reduced by 10% per warning range
-				float penalty = i < warnDist ? 0 : 0.1f * i / warnDist;
+				if (bestScore > 1f - penalty)
+					return false; //aint getting any better than this
 
-				score = mmForward.StepForward() - penalty;
-				if (score > bestScore) {
-					bestScore = score;
-					bestMatch = mmForward.Path();
+				// activate matchers as we enter their working range
+				while (active.end < mms.Length && mms[active.end].WorkingRange.Contains(loc))
+					active.end++;
+
+				// active MatchMatrix runs
+				for (int i = active.first; i <= active.last; i++) {
+					var mm = mms[i];
+					if (!mm.Match(loc, out float score)) {
+						Debug.Assert(i == active.first, "Match matricies out of order?");
+						active.first++;
+						continue;
+					}
+
+					if (penalty > 0) //ignore penalty for the first 10%
+						score -= penalty;
+
+					if (score > bestScore) {
+						bestScore = score;
+						bestMatch = mm.Path();
+					}
 				}
 
-				score = mmReverse.StepBackward() - penalty;
-				if (score > bestScore) {
-					bestScore = score;
-					bestMatch = mmReverse.Path();
-				}
+				loc += dir;
+				penalty += penaltyPerLine;
 
-				//aint getting any better than this
-				if (bestScore + penalty > 1f)
-					break;
+				return true;
 			}
-
-			return bestMatch;
 		}
 	}
 }
